backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-generativeai==0.5.4
Pillow==10.3.0
requests==2.31.0
python-dotenv==1.0.1
aiosqlite==0.20.0
apscheduler==3.10.4
websockets==12.0
aiofiles==23.2.1
